+++ src/assembly.rs
@@ -164,6 +164,7 @@
 //!
 //!
 
+pub mod buffers;
 pub mod global;
 pub mod local;
 pub mod operators;
